# Optimize `String#<<` with a single byte argument